refactor: move dropdown logic to Autocomplete component
